Steering Angle Prediction/requirements.txt
numpy==1.19.2
opencv-python==4.5.1.48
Pillow==8.1.0
torch==1.7.1
torchvision==0.8.2
tensorboard==2.4.1
tensorboardX==2.1
scikit-learn==0.23.2
pandas==1.2.1
tqdm==4.57.0
matplotlib==3.3.4
